feat - 자동차 move 기능을 전략으로서 추출
