+++ src/main/java/step3/Car.java
@@ -1,25 +1,15 @@
 package step3;
 
-import java.util.Random;
-
 public class Car {
-    private static final Integer bound = 9;
 
-    private static final Random random = new Random();
     private int location = 0;
 
+    private final MoveStrategy moveStrategy = new RandomStrategy();
 
     public void move() {
-        if (goOrStay(random.nextInt(bound))) {
+        if (moveStrategy.move()) {
             goForward();
         }
-    }
-
-    public boolean goOrStay(int key) {
-        if (key > bound || key < 0) {
-            throw new IllegalArgumentException(ErrorMessage.INVALID_RANDOM_VALUE);
-        }
-        return key > 4;
     }
 
     private void goForward() {
