+++ src/main/java/step3/MoveStrategy.java
@@ -1,0 +1,7 @@
+package step3;
+
+public interface MoveStrategy {
+
+    public boolean move();
+
+}
